feat(auto-refresh): move refresh logic out of getAccessToken and getIdToken
Both validate their tokens, but refreshing will be handled by middleware.
