+++ src/utils/getIdToken.ts
@@ -1,10 +1,10 @@
 import { config } from "../config";
 import { sessionManager } from "../session/sessionManager";
 import { NextApiRequest, NextApiResponse } from "next";
-import { validateToken } from "./validateToken";
+import { validateToken } from "./jwt/validation";
 import { kindeClient } from "../session/kindeServerClient";
 
-export const getIdToken = async (req: NextApiRequest, res: NextApiResponse) => {
+export const getIdToken = async (req: NextApiRequest, res?: NextApiResponse) => {
   const tokenKey = "id_token";
   try {
     const session = await sessionManager(req, res);
@@ -17,36 +17,16 @@ export const getIdToken = async (req: NextApiRequest, res: NextApiResponse) => {
       return null;
     }
 
-    // const isTokenValid = await validateToken({
-    //   token,
-    // });
+    const isTokenValid = await validateToken({
+      token,
+    });
 
-    // if (!isTokenValid) {
-    //   try {
-    //     const refreshSuccess = await kindeClient.refreshTokens(session);
-    //     if (refreshSuccess) {
-    //       const newToken = await session.getSessionItem(tokenKey);
-    //       const isNewTokenValid = await validateToken({
-    //         token: newToken as string,
-    //       });
-    //       if (isNewTokenValid) {
-    //         return newToken;
-    //       }
-    //     }
-    //     if (config.isDebugMode) {
-    //       console.error("getIdToken: token refresh failed");
-    //     }
-    //   } catch (error) {
-    //     if (config.isDebugMode) {
-    //       console.error("getIdToken: error during token refresh", error);
-    //     }
-    //   }
-
-    //   if (config.isDebugMode) {
-    //     console.error("getIdToken: invalid token");
-    //   }
-    //   return null;
-    // }
+    if (!isTokenValid) {
+      if (config.isDebugMode) {
+        console.error("getIdToken: invalid token");
+      }
+      return null;
+    }
 
     return token;
   } catch (error) {
